+++ package.json
@@ -24,6 +24,6 @@
     "posthtml-include": "^1.7.4"
   },
   "dependencies": {
-    "axios": "^1.3.5"
+    "axios": "^1.3.6"
   }
 }
