+++ src/partials/gallery.html
@@ -2,5 +2,10 @@
   <div class="movie-section">
     <ul class="movie-section__card"></ul>
   </div>
+  <div class="pagination">
+    <div class="page" id="prev">Previous Page</div>
+    <div class="current" id="current">1</div>
+    <div class="page" id="next">Next Page</div>
+</div>
 </section>
 <script src="/src/js/keyword.js" type="module"></script>
