+++ src/partials/header.html
@@ -37,4 +37,9 @@
   <div class="container">
     <ul class="Gallery GalleryCointainer" id="gallery-main"></ul>
   </div>
+<div class="pagination">
+    <div class="page" id="prevSearch">Previous Page</div>
+    <div class="current" id="currentSearch">1</div>
+    <div class="page" id="nextSearch">Next Page</div>
+</div>
 </div>
