+++ src/partials/loader.html
@@ -1,1 +1,1 @@
-<div class="loader"></div>
+<!-- <div class="loader"></div> -->
