+++ src/partials/pagination.html
@@ -1,5 +1,5 @@
-<div class="pagination">
+<!-- <div class="pagination">
     <div class="page" id="prev">Previous Page</div>
     <div class="current" id="current">1</div>
     <div class="page" id="next">Next Page</div>
-</div>
+</div> -->
